Refactor PCMNodeCalculator
